[Aaron Ballman] Hello!  I noticed there was a "fixme" listed in CrashRecoveryContext.cpp and that some Win32 support was needed.  I believe this patch will solve the issue.  I have tested it with Visual Studio 2010, and MinGW gcc 4.5.2 and the behavior is the same as what I get on OS X.
I was unable to get the test suite to run in my environment (had all
the prerequisites installed, followed the instructions on the site),
so I've not been able to run this against the suite.

One thing to note about this patch is that is only provides support
for Windows XP and higher.  Trying to support older versions of
Windows is possible (Win2k and higher) but comes with more pitfalls
(as noted in the comments).

If you have any questions, please ask!